Added sprite update and velocity

+++ src/main/java/fr/proneus/engine/input/MousePosition.java
@@ -3,28 +3,27 @@ package fr.proneus.engine.input;
 import fr.proneus.engine.Game;
 
 public class MousePosition {
-	
-	private Game game;
-	private int x, y;
-	
-	public MousePosition(Game game, int x, int y) {
-		this.game = game;
-		this.x = x;
-		this.y = y;
-	}
-	
-	public int getX() {
-		return x;
-	}
-	
-	public int getY() {
-		return y;
-	}
-	
-	public MousePosition toCameraPosition(){
-		return new MousePosition(game, x-(int)game.getCamera().getX(), y-(int)game.getCamera().getY());
-	}
-	
-	
+
+    private Game game;
+    private float x, y;
+
+    public MousePosition(Game game, float x, float y) {
+        this.game = game;
+        this.x = x;
+        this.y = y;
+    }
+
+    public float getX() {
+        return x;
+    }
+
+    public float getY() {
+        return y;
+    }
+
+    public MousePosition toCameraPosition() {
+        return new MousePosition(game, x - game.getCamera().getX(), y - game.getCamera().getY());
+    }
+
 
 }
